refactor(popover): introduce popover interaction kind type
