+++ packages/popover/src/index.ts
@@ -1,3 +1,4 @@
 export type { Placement as PopoverPlacement } from '@floating-ui/dom';
 export type { PopoverProps, Offset } from './Popover';
-export { Popover, PopoverInteractionKind } from './Popover';
+export { Popover } from './Popover';
+export { PopoverInteractionKind } from './types';
