+++ packages/popover/src/types.ts
@@ -1,0 +1,8 @@
+enum PopoverInteractionKind {
+  HOVER = 'hover',
+  HOVER_TARGET_ONLY = 'hover-target-only',
+  HOVER_OR_FOCUS = 'hover-or-focus',
+  CLICK = 'click',
+}
+
+export { PopoverInteractionKind };
